Stage error handling
  - Changed several panic errors to warnings.  Failure to execute
    a command is communicated via return value.  Only changed
    errors I came across.

    - several changes due to downstream error handling had to be
      made

 - FileSystemWatcher won't accept the default config and this
   induced an error elsewhere.  This was fixed.

+++ src/algo/unwarp.cu
@@ -196,7 +196,7 @@ extern "C" int unwarp_gpu(Array *out, Array *in, float duty)
       uint width_in  =  in->dims[0],
            width_out = out->dims[0];
       dim3 threads(BLOCK_SIZE,1,1),
-           grid(ceil(nrows/(float)BLOCK_SIZE),1,1);
+           grid((unsigned int)ceil(nrows/(float)BLOCK_SIZE),1,1);
 
       switch(in->type)
       {
